backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
ultralytics>=8.2.0
pillow>=10.4.0
python-multipart>=0.0.6
opencv-python>=4.10.0.84
numpy>=1.26.0
pydantic>=2.5.0
aiofiles>=23.2.1
moviepy>=1.0.3
